Move CLI out of the project.
