+++ CMakeLists.txt
@@ -2,5 +2,6 @@ cmake_minimum_required(VERSION 3.7)
 project(pddl_solver)
 add_compile_options(-Wall -Wextra)
 set(CMAKE_CXX_STANDARD 11)
+add_subdirectory(CLI)
 add_subdirectory(Parser)
 add_subdirectory(Planner)
