+++ Planner/CMakeLists.txt
@@ -6,6 +6,7 @@ add_compile_options(-Wall -Wextra)
 set(HEADER_DIR include)
 set(SOURCE_DIR src)
 set(PARSER_HEADER_DIR ../Parser/include)
+set(CLI_DIR ../CLI)
 
 file(GLOB SOURCE_FILES ${SOURCE_DIR}/*)
 file(GLOB HEADER_FILES ${HEADER_DIR}/*)
@@ -13,6 +14,7 @@ file(GLOB HEADER_FILES ${HEADER_DIR}/*)
 include_directories(${HEADER_DIR})
 include_directories(${SOURCE_DIR})
 include_directories(${PARSER_HEADER_DIR})
+include_directories(${CLI_DIR})
 include_directories(${CMAKE_CURRENT_BINARY_DIR})
 
 add_executable(
@@ -23,5 +25,6 @@ add_executable(
 
 target_link_libraries(
         ${CMAKE_PROJECT_NAME}
+        CLI11
         PDDLParser
 )
